[prism] tage and associated code to it's own file. typo fixes
